# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chip Quality Platform API
  description: |
    # Manufacturing Quality Inspection Platform API
    
    Production-grade REST API for semiconductor manufacturing quality control workflows with ML-powered defect detection.
    
    ## 🚀 Key Features
    - **🔄 Idempotency**: All mutations support Stripe-style idempotency keys with 24h retention
    - **📖 Pagination**: Cursor-based pagination for stable, consistent results
    - **🔐 Security**: OAuth2 + JWT with fine-grained scope-based authorization
    - **⚡ Async Processing**: Non-blocking inference with real-time event updates
    - **📊 Observability**: Comprehensive tracing, metrics, and structured logging
    - **🛡️ Rate Limiting**: Intelligent rate limiting with token bucket algorithm
    - **🔧 Resilience**: Circuit breakers, retries, and graceful degradation
    
    ## 🔐 Authentication & Authorization
    
    ### Token-Based Authentication
    All endpoints (except `/v1/token`) require a JWT bearer token:
    ```http
    Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
    ```
    
    ### OAuth2 Scopes
    The API uses fine-grained scopes for authorization:
    - `inspections:read` - View inspection data
    - `inspections:write` - Create and modify inspections
    - `defects:read` - View defect information
    - `defects:write` - Modify defect data
    - `reports:read` - Access reports and analytics
    - `reports:write` - Generate custom reports
...
    
...
    
...
    
...
    ```http
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /inspections:
    post:
      summary: Create new inspection
      description: |
        Creates a new inspection record for quality control workflow.
        This endpoint is idempotent when used with the `Idempotency-Key` header.
        
        **Typical Workflow**:
        1. Create inspection (this endpoint)
        2. Upload artifacts → `POST /inspections/{id}/artifacts`
        3. Trigger inference → `POST /inspections/{id}/infer`
        4. Check results → `GET /inspections/{id}`
        5. Generate report → `POST /reports`
        
        **Business Rules**:
        - Each (lot_id, chip_id, inspection_type) combination must be unique
        - Inspection priority affects ML inference queue position
        - Metadata can store custom key-value pairs for workflow integration
      operationId: createInspection
      tags:
        - Inspections
      security:
        - bearerAuth: [inspections:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionCreate'
            examples:
              visual_inspection:
                summary: Visual PCB Inspection
                value:
                  lot_id: "LOT-2025-001"
                  chip_id: "PCB-A1234"
                  inspection_type: "visual"
                  priority: 5
                  metadata:
                    station_id: "STATION-01"
                    operator: "john.doe"
                    shift: "morning"
              electrical_test:
                summary: Electrical Testing
                value:
                  lot_id: "LOT-2025-002"
                  chip_id: "CHIP-B5678"
                  inspection_type: "electrical"
                  priority: 8
                  metadata:
                    test_program: "ELECTRICAL_V2.1"
                    voltage_level: "3.3V"
      responses:
        '201':
          description: Inspection created successfully
          headers:
            Location:
              description: URL of the created inspection
              schema:
                type: string
                format: uri
                example: "/v1/inspections/550e8400-e29b-41d4-a716-446655440000"
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '200':
          description: Idempotent response - inspection already exists with same data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/IdempotencyConflict'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List inspections with filtering and pagination
      description: |
        Retrieve a paginated list of inspections with optional filtering.
        Uses cursor-based pagination for consistent results.
        
        **Filtering Options**:
        - By status (pending, in_progress, completed, failed)
        - By chip_id (exact match)
        - By lot_id (exact match)
        - By inspection_type
        - By date range (created_at)
        
        **Sorting**: Results are sorted by created_at DESC for consistent pagination.
      operationId: listInspections
      tags:
        - Inspections
      security:
        - bearerAuth: [inspections:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CorrelationId'
        - name: status
          in: query
          description: Filter by inspection status
          schema:
            $ref: '#/components/schemas/InspectionStatus'
          example: pending
        - name: chip_id
          in: query
          description: Filter by chip ID (exact match)
          schema:
            type: string
            maxLength: 100
          example: "PCB-A1234"
        - name: lot_id
          in: query
          description: Filter by lot ID (exact match)
          schema:
            type: string
            maxLength: 100
          example: "LOT-2025-001"
        - name: inspection_type
          in: query
          description: Filter by inspection type
          schema:
            $ref: '#/components/schemas/InspectionType'
          example: visual
        - name: created_after
          in: query
          description: Filter inspections created after this timestamp
          schema:
            type: string
            format: date-time
          example: "2025-10-29T00:00:00Z"
        - name: created_before
          in: query
          description: Filter inspections created before this timestamp
          schema:
            type: string
            format: date-time
          example: "2025-10-29T23:59:59Z"
      responses:
        '200':
          description: Paginated list of inspections
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Inspection'
  /inspections/{inspection_id}:
    get:
      summary: Get inspection details
      description: |
        Retrieve detailed information about a specific inspection,
        including current status, metadata, and associated defects.
      operationId: getInspection
      tags:
        - Inspections
      security:
        - bearerAuth: [inspections:read]
      parameters:
        - $ref: '#/components/parameters/InspectionId'
        - $ref: '#/components/parameters/CorrelationId'
        - name: include_defects
          in: query
          description: Include associated defects in response
          schema:
            type: boolean
            default: false
          example: true
      responses:
        '200':
          description: Inspection details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionDetailed'
...
  /reports:
    post:
      summary: Generate custom report
      description: |
        Generate analytics reports with custom parameters and filters.
        Supports multiple output formats and scheduled delivery.
        
        **Report Types**:
        - `defect_summary` - Aggregated defect statistics
        - `inspection_throughput` - Production metrics over time  
        - `quality_trends` - Quality KPIs and trends
        - `detailed_analysis` - Comprehensive inspection details
      operationId: generateReport
      tags:
        - Reports
      security:
        - bearerAuth: [reports:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report generation queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'

  # ===== OPERATIONS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /token endpoint.
        Include in Authorization header: `Bearer <token>`
  parameters:
    InspectionId:
      name: inspection_id
      in: path
      required: true
      description: Unique inspection identifier
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Unique key for idempotent operations. Recommended format:
        `{user_id}_{timestamp}_{operation}` (max 255 chars)
      schema:
        type: string
        maxLength: 255
        pattern: '^[a-zA-Z0-9_-]+$'
      example: "user123_2025-10-29T10:30:00Z_create-inspection"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Request correlation ID for distributed tracing
      schema:
        type: string
        format: uuid
      example: "01234567-89ab-cdef-0123-456789abcdef"
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Pagination cursor from previous response.
        Opaque string - do not attempt to decode.
      schema:
        type: string
        maxLength: 1000
      example: "eyJjcmVhdGVkX2F0IjoiMjAyNS0xMC0yOVQxMDozMDowMFoifQ"
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
  headers:
    X-Request-ID:
      description: Unique request identifier for tracking and support
      schema:
        type: string
        pattern: '^req_[a-zA-Z0-9]{16}$'
      example: "req_1234567890abcdef"
  schemas:
...
    InspectionStatus:
      type: string
      enum: [pending, in_progress, completed, failed]
      description: Current inspection status
      example: pending
    InspectionType:
      type: string
      enum: [visual, electrical, thermal, dimensional, functional]
      description: Type of quality inspection
      example: visual
...
    InspectionCreate:
      type: object
      required:
        - lot_id
        - chip_id
        - inspection_type
      properties:
        lot_id:
          type: string
          description: Manufacturing lot identifier
          maxLength: 100
          pattern: '^[A-Z0-9-]+$'
          example: "LOT-2025-001"
        chip_id:
          type: string
          description: Individual chip/component identifier
          maxLength: 100
          pattern: '^[A-Za-z0-9-]+$'
          example: "PCB-A1234"
        inspection_type:
          $ref: '#/components/schemas/InspectionType'
        priority:
          type: integer
          description: Processing priority (1=lowest, 10=highest)
          minimum: 1
          maximum: 10
          default: 5
          example: 5
        metadata:
          type: object
          description: Custom key-value pairs for workflow integration
          additionalProperties:
            type: string
          example:
            station_id: "STATION-01"
            operator: "john.doe"
            shift: "morning"
    Inspection:
      allOf:
        - $ref: '#/components/schemas/InspectionCreate'
        - type: object
          required:
            - id
            - status
            - created_at
            - updated_at
          properties:
            id:
              type: string
              format: uuid
              description: Unique inspection identifier
              readOnly: true
              example: "550e8400-e29b-41d4-a716-446655440000"
            status:
              $ref: '#/components/schemas/InspectionStatus'
            created_at:
              type: string
              format: date-time
              description: Creation timestamp
              readOnly: true
              example: "2025-10-29T10:30:00.000Z"
            updated_at:
              type: string
              format: date-time
              description: Last modification timestamp
              readOnly: true
              example: "2025-10-29T10:35:00.000Z"
    InspectionDetailed:
      allOf:
        - $ref: '#/components/schemas/Inspection'
        - type: object
          properties:
            defects:
              type: array
              description: Associated defects (if include_defects=true)
              items:
                $ref: '#/components/schemas/Defect'

    # ===== DEFECTS =====
...
    ReportRequest:
      type: object
      required:
        - report_type
        - title
        - date_range
        - format
      properties:
        report_type:
          type: string
          enum: [defect_summary, inspection_throughput, quality_trends, detailed_analysis]
          description: Type of report to generate
          example: "defect_summary"
        title:
          type: string
          description: Report title
          maxLength: 200
          example: "October 2025 Defect Analysis"
        date_range:
          type: object
          required: [start, end]
          properties:
            start:
              type: string
              format: date-time
              description: Start of reporting period
              example: "2025-10-01T00:00:00Z"
            end:
              type: string
              format: date-time
              description: End of reporting period
              example: "2025-10-31T23:59:59Z"
        format:
          type: string
          enum: [pdf, excel, csv, json]
          description: Output format
          example: "pdf"
    ReportResponse:
      type: object
      required:
        - operation_id
        - status
        - estimated_completion
      properties:
        operation_id:
          type: string
          pattern: '^op_[a-zA-Z0-9]{16}$'
          description: Operation ID for tracking report generation
          example: "op_report_123456789"
        status:
          type: string
          enum: [queued, processing, completed, failed]
          description: Generation status
          example: "queued"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-10-29T10:33:00.000Z"

    # ===== OPERATIONS =====
...
    PaginatedResponse:
      type: object
      required:
        - data
        - has_more
      properties:
        data:
          type: array
          description: Result data
          items: {}
        has_more:
          type: boolean
          description: Whether more results are available
          example: true
        next_cursor:
          type: string
          description: Cursor for next page
          nullable: true
          example: "eyJjcmVhdGVkX2F0IjoiMjAyNS0xMC0yOVQxMDozMDowMFoifQ"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "validation_error"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              description: Additional error context
              nullable: true
              additionalProperties: true
              example:
                field: "chip_id"
                constraint: "must be alphanumeric, 3-50 characters"
            request_id:
              type: string
              description: Request ID for tracking
              example: "req_1234567890abcdef"
            timestamp:
              type: string
              format: date-time
              description: Error occurrence time
              example: "2025-10-29T10:30:00.000Z"
  responses:
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdempotencyConflict:
      description: Idempotency key reused with different payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'